shill: WiFi: Timeout pending connections
Whenever we select a pending service, start a timeout, at
the end of which we will disconnect and fail the connection.
There are a one cleanup associated with this -- prevent
SelectService(NULL) from being called when current_service_
becomes NULL but the currently selected service is
pending_service_.

BUG=chromium-os:23206
TEST=New unit test.  Connect to non-existent network using
"Join Other ...", and observe failure mole.

Change-Id: I1e4954c4bef55406b8fdbf0d6edb352163886f12
Reviewed-on: https://gerrit.chromium.org/gerrit/27442
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>