Enable/disable battery polling on resume/suspend
Curently the battery polling code is ignorant to suspend/resume
events. This means that on resume the next battery update is anywhere
within 30 seconds, so the UX is displaying wrong values. Compounding
this is the fact that if the timer fires within the first 5 seconds
after resuming the daemon will send up values with an artificially low
discharge rate.

To solve this issue the code has been changed to halt polling on suspend
and restart it on resume. This requires expanding the interface between
the suspender and the daemon.

BUG=chromium-os:32919
TEST=Built power_manager and ran unit tests.
     Build image and ran on device.
     Checked that power status in the UI went to "Calculating" on resume
     then after ~5s goes to the correct value.

Change-Id: Ia44db9c8d3673540e5ffbe41c09f3516ca394b4a
Reviewed-on: https://gerrit.chromium.org/gerrit/28396
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>