powerd: extend audio threshold time to 5 seconds
The audio poll time and the audio activity threshold are both 1 second.
The calculated audio time-since-last should be in the range [0, 1 sec]
but it sometimes exceeds 1 second by a little bit.  (e.g. 1050 ms).
This is interpreted as "audio is no longer playing" since it exceeds the
threshold.

The fundamental problem is that the poll interval and threshold time are
the same.  They need to be significantly different.  The solution is to
extend the threshold time to 5 seconds.

BUG=chromium-os:32327
TEST=see BUG

Change-Id: Ie3c4da6fdf9fdfac822d24b55e4514811b73a279
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28561
Reviewed-by: Benson Leung <bleung@chromium.org>