Don't suspend if headphones are attached to stumpy.
Based on a new use flag, stay_awake_with_headphones, change the suspend
behavior.  Don't suspend if there are headphones attached and the use
flag is set.  If powered speakers are plugged in and Stumpy suspends, a
loud hiss is generated due to the ground pin of the headphone jack
floating.  This workaround avoids the annoying sound at the cost of
always keeping stumpy awake if anything is plugged in to the front
headphone jack.

CQ-DEPEND=CL:30827 CL:29329

BUG=chrome-os-partner:10413
TEST=Leave stumpy idle while headphones attached, make sure it doesn't
suspend.  Do the same for lumpy, checking that it still suspends.  Also
check that Lumpy's lid still suspends.

Change-Id: I8d9988c1a301dbf5f3b1fc1fa868f641cec345e0
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28603
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>