Create an AIK and assemble parameters required for remote attestation.
An Attestation Identity Key (AIK) is required for remote attestation. It
allows us to quote PCRs and certify other keys. Creating an AIK requires
TPM owner privilege and so this work must be done before discarding the
owner password. At the time we generate the AIK we collect all
parameters that we need to verify it's authenticity and keep them until
we have a chance to talk to a CA which can validate and certify the AIK
based on these parameters.

BUG=chromium-os:32772
TEST=All unit tests pass.
     Manually verified that the epb file exists after initial login and
     can be parsed and decrypted.

Change-Id: I7d951a52c497512207c1757d2e22e22c0b9f04d3
Reviewed-on: https://gerrit.chromium.org/gerrit/29101
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Commit-Ready: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>