Raise the ui throttling limit to 10 times per min
It's easily possible for automated tests to cross the current limit
(6 times / min) as a part of a valid use case.

Raising it to 10, which is a bit more unrealistic, except in real crashing
scenarios which is where the throttle should really apply.

I've verified that crashing 10 times does not make a slow machine unusable
(verified on alex)

BUG=chromium-os:33165
TEST=None

Change-Id: If47f968765501f38ab5272177d55b800d8194fea
Reviewed-on: https://gerrit.chromium.org/gerrit/28961
Commit-Ready: Nirnimesh <nirnimesh@chromium.org>
Reviewed-by: Nirnimesh <nirnimesh@chromium.org>
Tested-by: Nirnimesh <nirnimesh@chromium.org>