init: Defer starting X until boot splash animation is done.
This adds a new boot-splash task for running ply-image and
makes the ui job wait for boot-splash to stop before
starting.  It also adds a boot-alert-ready abstract job that
runs once boot-splash is done and makes the
chromeos-boot-alert script wait for boot-alert-ready.

I'm also increasing the boot animation speed to 40 FPS
(previously, it was 20), which cuts the the time between
the pre-startup and x-started boot events by 10-70 ms on my
Lumpy dev build.

BUG=chromium-os:32667
TEST=manual: checked that the normal path still works; added an extra boot-alert call to chromeos_startup and checked that it runs

Change-Id: Ib22b533ceb4af929a66e781546ee01894ad390d0
Reviewed-on: https://gerrit.chromium.org/gerrit/28999
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Ready: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>