shill: Retry calls to proxy's GetIMSI
After unlocking a SIM, it takes some time for the SIM to be
ready to respond to the GetIMSI request.  The amount of time
is unknown.  This change adds retry logic to retry the GetIMSI
request every 200 ms up to 10 times.

BUG=chromium-os:31866
TEST=On a lumpy, unlock SIM and verify that modem becomes enabled.
Run all unit tests.

Change-Id: I7c9fea3f14e05aa9873d86eaf9221fd121ddb4ac
Reviewed-on: https://gerrit.chromium.org/gerrit/29120
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Gary Morain <gmorain@chromium.org>
Tested-by: Gary Morain <gmorain@chromium.org>