handle interrupted Lockbox creation
It is easy to race the Lockbox creation on Snow, so make sure cryptohome
attempts to create the Lockbox where possible. (i.e. since TPM ownership
may have happened at a prior boot, the flag for "first boot" isn't valid
anymore, but the TPM password is still around because it has not
finalized.)

BUG=chromium-os:33218
TEST=snow build, manual testing, unit tests pass

Change-Id: I27ead3cac0ff23b327eb512cef18c3033a01776a
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29122
Reviewed-by: Darren Krahn <dkrahn@chromium.org>