shill: IPConfig: Create DBus API method "Refresh"
Calling this method will refresh the IPConfig state.  This means
in the case of a DHCP IPConfig, the lease will be renewed and
any changed StaticIP configuration for the service will be
newly applied.  Also clean up MoveBefore and MoveAfter methods
which are neither in the docs, supplied in the service constants
or used in the Chrome libraries.

BUG=chromium-os:33223
TEST=Manual: Call method on shill and watch for DHCP messages.

Change-Id: I7f44c94e9c76993f93d1a5cf95eafa0be63da957
Reviewed-on: https://gerrit.chromium.org/gerrit/29182
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>