clobber: shred rotational drives instead of full wipe
In an attempt to provide a faster wipe of larger disks, perform a
shred of unencrypted HDD contents, and then force the wipe into "fast"
mode. Since the encrypted partition key and eCryptFS keys are needed
to recover the encrypted portions of the drive, shredding the keys is
sufficient on rotational disks (since we do not use "data=journal" when
mounting the ext4 filesystem). On SSD, a full zeroing is still required
to be sure the files have actually been deleted from the NANDs.

Adding TRIM calls for SSDs will be part of a separate CL, since the HDD
logic needs to be explicitly backported to R21 for Parrot.

BUG=chrome-os-partner:11342
TEST=parrot build, manual testing

Change-Id: Ic41956cf5308538dccdc6f43d3c09089849e2f63
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29365
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>