fix -I flag paths when building out-of-tree
The current build logic assumes the build dir is directly under the
source dir.  This fails if the build and source trees are completely
different paths.  So use full paths to the source/build dirs rather
than relative ones so that it works in all states.

BUG=chromium-os:33327
TEST=`cros_run_unit_tests --board x86-alex -p shill` still works

Change-Id: I8d03ca45a5b84fec12980dac6d427b0868e154c4
Reviewed-on: https://gerrit.chromium.org/gerrit/29385
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>