powerd_resume: made hwclock be cached from disk before suspend
I added the line '/sbin/hwclock > /dev/null' which caches the binary before
suspending (when the TEST_flag is set to true).

I also capture the output of hwclock after resume then echo that to the file to
make sure we are not waiting on the disk before getting the time.

BUG=chrome-os-partner:11780
TEST=power_Resume, powerd_suspend, opening and closing laptop

Change-Id: I0658033fbc6f788f99fafe808c714a6a5d199b6c
Reviewed-on: https://gerrit.chromium.org/gerrit/29538
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>