shill: Change cellular callback bindings to weak pointers.
Change to using weak pointers in DBus callback bindings so that we do
not hold a reference count out on the object while calling into modem
manager.  This becomes a problem if an async DBus call is in progress
and we release the proxy.  The callback does not get executed which
leaves a dangling reference count and keeps the object alive.  This stale
DBus object prevents us from properly registering the new DBus object
since the two paths are the same.

BUG=chromium-os:31636
TEST=Unit tests, network_3GSuspendResume

Change-Id: I252f9e28978b5426509bfafdd66adbfd8b289936
Reviewed-on: https://gerrit.chromium.org/gerrit/29667
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Thieu Le <thieule@chromium.org>
Commit-Ready: Thieu Le <thieule@chromium.org>