shill: WiFi: Mark pending service idle on disconnect
When disconnecting from the pending service, we cannot depend on
wpa_supplicant to set the CurrentBSS in order to drive the service
state to "idle".  This case was covered when connecting to a new
service while there was already a pending service, but this issue
can happen in other situations, such as when all the pending
service's BSSes are removed while connecting, or any other situation
where DisconnectFrom is called on the pending service.

BUG=chrome-os-partner:12111,chromium-os:33512
TEST=network_WiFiRoaming.010RetryReconnectHidden, new unit test

Change-Id: I77e4fcf52ea6b2ad67bc21f11fb2786ee9409c1a
Reviewed-on: https://gerrit.chromium.org/gerrit/30080
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>