shill: Dump MemoryLog to disk on disconnects
A good starting place to dump verbose shill logs is when we disconnect
from a service (like wifi).  Add a minute amount of boiler plate to
MemoryLog to make dumping simple, and some more unit tests for that logic.

In order to keep our logs private, we store them to

  /home/chronos/user/log/connectivity.log

when logged in, and

  /var/log/connectivity.log

when logged out.  We use the presence of /var/run/state/logged-in to
detect our current state.  In order to keep a user from logging
something interesting, then logging out and putting that string in the
logged out logs, we clear the MemoryLog when we receive a PopProfile
signal from the logout script.

In addition to a unit test, confirmed by hand cycling a lumpy on and off
the wireless that log files are indeed generated and rotated as
expected.

TEST=Unit tests pass, added new tests, hand testing as described above
BUG=chromium-os:33514

Change-Id: Ib1e08217f920c47289f30ab5d0f4f198f06c9e82
Reviewed-on: https://gerrit.chromium.org/gerrit/30477
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Christopher Wiley <wiley@chromium.org>