network_diagnostics: harmonize diag_link_wifi and diag_wifi
Change diag_link_wifi to call diag_wifi for wifi statistics, instead of
capturing a subset of the diag_wifi information on its own. With this
change, we can ask users to just run "network_diag" in crosh, instead of
both "network_diag" (for general diagnosis) and "network_diag --wifi" (for
wifi-specific data).

The down-side is that we gather info for all wifi devices instead of just
the failed one. But multiple devices should not be a common case.

BUG=chromium-os:33558
TEST=manual (see below)

Manual testing
1. connect to GoogleGuest
2. iptables -I INPUT -i wlan0 -j DROP
3. run network_diagnostics
4. observe that output includes contents of files like
   /sys/kernel/debug/ieee80211/phy0/ath9k/dma

Change-Id: I20e043846ec3a179a6fef330c30b2cb994a3c40f
Reviewed-on: https://gerrit.chromium.org/gerrit/30437
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>