powerd - Detect failed writes to /sys/power/state
If a driver fails to suspend, the error code is returned from the
write call to /sys/power/state. Use this to detect if suspend fails,
and send that status along with the PowerStateChanged dbus message.

If suspend fails, powerman shouldn't reset the retry count.

BUG=chromium-os:33675
TEST=modified tpm driver to always fail suspend; try suspend
Change-Id: I25796648328b86356b9412f244d0b91dfc9b3ab1
Reviewed-on: https://gerrit.chromium.org/gerrit/30910
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Commit-Ready: Michael Spang <spang@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>