Change initial start of polling battery poling to a resume.
This causes the daemon to poll the state of battery on start and signal the
browser that there is information about the battery state to pull. This also
schedules the next poll in 5 seconds since the time values from the first poll
might be bogus and need to be ignored. This is the same behaviour that is used
when the device transitions power state.

BUG=chromium:136885
TEST=Started device and watch the status tray on the login screen. The battery
     icon should appear in about of second, instead of taking upto 30 seconds,
     with the Calculating estimate and then after 5 seconds should transition to
     the correct time estimate.

Change-Id: I948dcfe467e4d7172f8ed57ba8437cf767dd6a52
Reviewed-on: https://gerrit.chromium.org/gerrit/31140
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>