Init: add upstart job to move attestation
We are moving attestation.epb to
/mnt/stateful_partition/unencrypted/preserve so that when we do a
stateful wipe it is preserved.

This upstart job will move the file for any systems that have the file
in the original directory.

This is currently the only file we need to preserve for consumers.

CQ-DEPEND=Iafd467e485fa16326d012c28847105094ebb0251
BUG=chromium-os:33357
TEST=logged in so that attestation.epb was created then rebooted and
ensured that the file was moved to the new location.

Change-Id: I5c24736ea95ccf8ecfa82032cb81ed126b98deb2
Reviewed-on: https://gerrit.chromium.org/gerrit/31504
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Simran Basi <sbasi@google.com>
Tested-by: Simran Basi <sbasi@google.com>
Commit-Ready: Simran Basi <sbasi@google.com>