Set minimum level when starting backlight controller
There is a requirement that the backlight be set to a minimum visible
level when starting powerd.  However, if the backlight is off, this
minimum visible level is not set.  The solution is to set to the minimum
visible level when starting backlight controller, which also turns on
the screen.

BUG=chrome-os-partner:12243
TEST=turn brightness all the way down. "restart powerd" over ssh. Screen
should be visible.

Change-Id: Ib95a4123e93989bad37a3290dfee524b0f0ac7f8
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31393
Reviewed-by: Daniel Erat <derat@chromium.org>