Instant transitions between min visible level and zero
Also create a pref flag to turn this feature on.  This allows for
transitions between 0 and |min_visible_level_| to be instantaneous
instead of gradual, for systems where we do not want to set the
backlight to anything inside that range.

BUG=chrome-os-partner:12622
TEST=echo something high, like 20% of max brightness, to
/var/lib/power_manager/min_visible_backlight_level.  echo 1 to
/var/lib/power_manager/instant_transitions_below_min_level
Restart powerd. Adjust brightness down to 0.  The screen should
directly turn off at the last step, without fading.

Change-Id: I675d1f201646435ef0ecd8dccbea0f996b166ef3
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31656
Reviewed-by: Daniel Erat <derat@chromium.org>